easylight/requirements.txt
diff-match-patch==20121119
Django==1.11.2
django-import-export==0.5.1
djangorestframework==3.6.3
et-xmlfile==1.0.1
jdcal==1.3
jsonfield==2.0.2
odfpy==1.3.5
openpyxl==2.4.8
psycopg2==2.7.1
Pygments==2.2.0
pytz==2017.2
PyYAML==3.12
tablib==0.11.5
unicodecsv==0.14.1
xlrd==1.0.0
xlwt==1.2.0
